[*]added possibility to make photo and check faces on photo

+++ skillary/SkillaryApp/Classes/ViewControllers/SKCaptureController.m
@@ -114,7 +114,7 @@
 }
 
 - (void)updateCounterLabel {
-    self.lbCounter.text = [NSString stringWithFormat:@"%d", currentCounterValiue];
+    self.lbCounter.text = [NSString stringWithFormat:@"%ld", (long)currentCounterValiue];
 }
 
 - (void)startCapturing {
